Changed category for configure subnetwork trace
