+++ src/Android/Xamarin.Android/Samples/Utility network/ConfigureSubnetworkTrace/ConfigureSubnetworkTrace.cs
@@ -25,7 +25,7 @@ namespace ArcGISRuntimeXamarin.Samples.ConfigureSubnetworkTrace
     [Activity(ConfigurationChanges = Android.Content.PM.ConfigChanges.Orientation | Android.Content.PM.ConfigChanges.ScreenSize)]
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
         "Configure subnetwork trace",
-        "Network analysis",
+        "Utility network",
         "Get a server-defined trace configuration for a given tier and modify its traversability scope, add new condition barriers and control what is included in the subnetwork trace result.",
         "")]
     public class ConfigureSubnetworkTrace : Activity
