+++ src/Forms/Shared/Samples/Utility network/ConfigureSubnetworkTrace/ConfigureSubnetworkTrace.xaml.cs
@@ -18,7 +18,7 @@ namespace ArcGISRuntimeXamarin.Samples.ConfigureSubnetworkTrace
 {
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
         "Configure subnetwork trace",
-        "Network analysis",
+        "Utility network",
         "Get a server-defined trace configuration for a given tier and modify its traversability scope, add new condition barriers and control what is included in the subnetwork trace result.",
         "")]
     public partial class ConfigureSubnetworkTrace : ContentPage
