+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Utility network/ConfigureSubnetworkTrace/ConfigureSubnetworkTrace.jpg
