+++ src/iOS/Xamarin.iOS/Samples/Utility network/ConfigureSubnetworkTrace/readme.metadata.json
@@ -1,5 +1,5 @@
 {
-    "category": "Network analysis",
+    "category": "Utility network",
     "description": "Get a server-defined trace configuration for a given tier and modify its traversability scope, add new condition barriers and control what is included in the subnetwork trace result.",
     "formal_name": "ConfigureSubnetworkTrace",
     "ignore": false,
@@ -18,11 +18,11 @@
         "validate consistency"
     ],
     "nuget_packages": {
-        "Esri.ArcGISRuntime": "100.7.0"
+        "Esri.ArcGISRuntime": "100.8.0"
     },
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/wpf/sample-code/configuresubnetworktrace.htm"
+        "/net/latest/ios/sample-code/configuresubnetworktrace.htm"
     ],
     "relevant_apis": [
         "CodedValueDomain",
@@ -49,8 +49,7 @@
         "UtilityTraversability"
     ],
     "snippets": [
-        "ConfigureSubnetworkTrace.xaml",
-        "ConfigureSubnetworkTrace.xaml.cs"
+        "ConfigureSubnetworkTrace.cs"
     ],
     "title": "Configure subnetwork trace"
 }
